anchor to new fork, split domain-specific commands to domain dirs (#23)
