+++ cmd/clone/cmd.go
@@ -1,10 +1,8 @@
-// Copyright © 2022 jesus m. rodriguez jmrodri@gmail.com
-//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -19,10 +17,10 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/jmrodri/gh2jira/internal/config"
-	"github.com/jmrodri/gh2jira/internal/gh"
-	"github.com/jmrodri/gh2jira/internal/jira"
-	"github.com/jmrodri/gh2jira/pkg/util"
+	"github.com/oceanc80/gh2jira/internal/config"
+	"github.com/oceanc80/gh2jira/internal/gh"
+	"github.com/oceanc80/gh2jira/internal/jira"
+	"github.com/oceanc80/gh2jira/pkg/util"
 )
 
 var (
