+++ internal/config/config_test.go
@@ -15,7 +15,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/jmrodri/gh2jira/pkg/util"
+	"github.com/oceanc80/gh2jira/pkg/util"
 	"github.com/stretchr/testify/require"
 )
