+++ main.go
@@ -1,10 +1,8 @@
-// Copyright © 2022 jesus m. rodriguez jmrodri@gmail.com
-//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,7 +15,7 @@ package main
 import (
 	"os"
 
-	"github.com/jmrodri/gh2jira/cmd/root"
+	"github.com/oceanc80/gh2jira/cmd/root"
 )
 
 // So we will want to allow this to be able to take in a specific GH issue id or
